# backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:8081",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/oti/": {
      "get": {
        "tags": [
          "Oti"
        ],
        "description": "Endpoint para obtener las otis.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "otis encontradas",
            "schema": {
              "$ref": "#/definitions/arrayOti"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
    "/oti/sectores": {
      "get": {
        "tags": [
          "Oti"
        ],
        "description": "Endpoint para obtener las otis asigandas de un sector.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "otis encontradas",
            "schema": {
              "$ref": "#/definitions/arraySector"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
    "/oti/{id}": {
      "get": {
        "tags": [
          "Oti"
        ],
        "description": "Endpoint para obtener una oti.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "oti ID"
          }
        ],
        "responses": {
          "200": {
            "description": "otis encontradas",
            "schema": {
              "$ref": "#/definitions/oti"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
    "/oti/obtenerRoscas/{idOrden}": {
      "get": {
        "tags": [
          "Oti"
        ],
        "description": "Endpoint para obtener las roscas de una orden.",
        "parameters": [
          {
            "name": "idOrden",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "orden ID"
          }
        ],
        "responses": {
          "200": {
            "description": "otis encontradas",
            "schema": {
              "$ref": "#/definitions/arrayRoscas"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
    "/oti/obtenerOrdenes/{fecha}": {
      "get": {
        "tags": [
          "Oti"
        ],
        "description": "Endpoint filtra ordenes por fecha.",
        "parameters": [
          {
            "name": "fecha",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "fecha creacion orden de produccion"
          }
        ],
        "responses": {
          "200": {
            "description": "otis encontradas",
            "schema": {
              "$ref": "#/definitions/arrayOrdenes"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
    "/oti/datosBasicos": {
      "post": {
        "tags": [
          "Oti"
        ],
        "description": "Endpoint para crear una oti con sus datos basicos.",
        "parameters": [
          {
            "name": "oti",
            "in": "body",
            "description": "datos para crear una oti",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/oti"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "se crea oti",
            "schema": {
              "$ref": "#/definitions/oti"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
    "/oti/sectoresYTareas": {
      "put": {
        "tags": [
          "Oti"
        ],
        "description": "Endpoint para actulizar sectores y tareas.",
        "parameters": [
          {
            "name": "oti",
            "in": "body",
            "description": "datos para actulizar una oti",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/oti"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "se actuliza oti",
            "schema": {
              "$ref": "#/definitions/oti"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
    "/oti/actualizarOrden": {
      "put": {
        "tags": [
          "Oti"
        ],
        "description": "Endpoint para actulizar orden ",
        "parameters": [
          {
            "name": "oti",
            "in": "body",
            "description": "datos para actulizar una oti",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/oti"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "se actuliza oti",
            "schema": {
              "$ref": "#/definitions/oti"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
...
    "/rosca/": {
      "get": {
        "tags": [
          "Rosca"
        ],
        "description": "Endpoint para obtener roscas .",
        "parameters": [],
        "responses": {
          "200": {
            "description": "se crea orden",
            "schema": {
              "$ref": "#/definitions/arrayRoscas"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      },
      "post": {
        "tags": [
          "Rosca"
        ],
        "description": "Endpoint para crear una rosca.",
        "parameters": [
          {
            "name": "rosca",
            "in": "body",
            "description": "datos para crear una rosca",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/rosca"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "se crea rosca",
            "schema": {
              "$ref": "#/definitions/rosca"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
    "/rosca/tipoDeRosca": {
      "get": {
        "tags": [
          "Rosca"
        ],
        "description": "Endpoint para obtener los tipos de roscas.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "se crea orden",
            "schema": {
              "$ref": "#/definitions/arrayTipoDeRosca"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
    "/rosca/{id}": {
      "get": {
        "tags": [
          "Rosca"
        ],
        "description": "Endpoint para obtener una rosca.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "rosca ID"
          }
        ],
        "responses": {
          "200": {
            "description": "se crea orden",
            "schema": {
              "$ref": "#/definitions/rosca"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
...
    "/sector/": {
      "get": {
        "tags": [
          "Sector"
        ],
        "description": "Endpoint para obtener todos los sectores.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "se obtiene sectores",
            "schema": {
              "$ref": "#/definitions/arraySector"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
    "/sector/{nombre}": {
      "get": {
        "tags": [
          "Sector"
        ],
        "description": "Endpoint para obtener un sector.",
        "parameters": [
          {
            "name": "nombre",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "nombre del sector"
          }
        ],
        "responses": {
          "200": {
            "description": "se obtiene sector",
            "schema": {
              "$ref": "#/definitions/sector"
            }
          },
          "400": {
            "description": "Error "
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "arraySector": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "613bc737b2153ee73d448bf1"
        },
        "nombre": {
          "type": "string",
          "example": "corte"
        },
        "activo": {
          "type": "boolean",
          "example": true
        },
        "operarios": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "613e631eb2153ee73d784e23"
              },
              "numeroEmpleado": {
                "type": "number",
                "example": 3
              },
              "DNI": {
                "type": "string",
                "example": "89-620-4885"
              },
              "nombre": {
                "type": "string",
                "example": "Fidelio"
              },
              "apellido": {
                "type": "string",
                "example": "Tedorenko"
              },
              "telefono": {
                "type": "string",
                "example": "332-243-1445"
              },
              "nombreUsuario": {
                "type": "string",
                "example": "ftedorenko2"
              },
              "contraseña": {
                "type": "string",
                "example": "eErDOWk7jQj0"
              },
              "email": {
                "type": "string",
                "example": "ftedorenko2@ucoz.ru"
              },
              "sector": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "corte"
                  }
                }
              }
            }
          }
        }
      }
    },
    "rosca": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "61a161477146661f9d634386"
        },
        "descripcionTecnica": {
          "type": "string",
          "example": "con corte especial en los ultimos 3/4 , para perforaciones horizontales y verticales"
        },
        "medida": {
          "type": "string",
          "example": "asd"
        },
        "tipoDeRosca": {
          "type": "object",
          "properties": {
            "descripcion": {
              "type": "string",
              "example": ""
            },
            "nombre": {
              "type": "string",
              "example": "TDL"
            },
            "categoria": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "premium"
                },
                "descripcion": {
                  "type": "string",
                  "example": "exportacion marina y utilizacion para caños de gas"
                },
                "_id": {
                  "type": "string",
                  "example": "61a1611f7146661f9d634384"
                }
              }
            },
            "_id": {
              "type": "string",
              "example": "613e604acb4275f300506790"
            }
          }
        }
      }
    },
    "arrayTipoDeRosca": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "613e604acb4275f300506790"
          },
          "descripcion": {
            "type": "string",
            "example": "descripcion_5"
          },
          "nombre": {
            "type": "string",
            "example": "TDL"
          },
          "categoria": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "61a391f17d3c127a35ddfa41"
              },
              "nombre": {
                "type": "string",
                "example": "premium"
              },
              "descripcion": {
                "type": "string",
                "example": "exportacion marina y utilizacion para caños de gas"
              }
            }
          }
        }
      }
    },
    "arrayRoscas": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "61a161477146661f9d634386"
          },
          "descripcionTecnica": {
            "type": "string",
            "example": "con corte especial en los ultimos 3/4 , para perforaciones horizontales y verticales"
          },
          "medida": {
            "type": "string",
            "example": "asd"
          },
          "tipoDeRosca": {
            "type": "object",
            "properties": {
              "descripcion": {
                "type": "string",
                "example": ""
              },
              "nombre": {
                "type": "string",
                "example": "TDL"
              },
              "categoria": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "premium"
                  },
                  "descripcion": {
                    "type": "string",
                    "example": "exportacion marina y utilizacion para caños de gas"
                  },
                  "_id": {
                    "type": "string",
                    "example": "61a1611f7146661f9d634384"
                  }
                }
              },
              "_id": {
                "type": "string",
                "example": "613e604acb4275f300506790"
              }
            }
          }
        }
      }
    },
...
    "arrayOti": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "613e5777b2153ee73d73142f"
          },
          "fechaFin": {
            "type": "string",
            "example": ""
          },
          "rosca": {
            "type": "object",
            "properties": {
              "descripcionTecnica": {
                "type": "string",
                "example": "es una rosca dificil de hacer"
              },
              "medida": {
                "type": "string",
                "example": "20x50x10"
              },
              "tipoDeRosca": {
                "type": "object",
                "properties": {
                  "descripcion": {
                    "type": "string",
                    "example": "la mejor rosca"
                  },
                  "categoria": {
                    "type": "object",
                    "properties": {
                      "nombre": {
                        "type": "string",
                        "example": "premium"
                      },
                      "descripcion": {
                        "type": "string",
                        "example": "exportacion marina y utilizacion para caños de gas"
                      },
                      "_id": {
                        "type": "string",
                        "example": "615e53c8db8fc5cc55e1fb3d"
                      }
                    }
                  },
                  "_id": {
                    "type": "string",
                    "example": "615e53c8db8fc5cc55e1fb3c"
                  }
                }
              },
              "_id": {
                "type": "string",
                "example": "615e53c8db8fc5cc55e1fb3b"
              }
            }
          },
          "sector": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "operario": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "_id": {
                  "type": "string",
                  "example": "61a382527d3c127a35ddf813"
                },
                "nombre": {
                  "type": "string",
                  "example": "deposito"
                }
              }
            }
          },
          "tareas": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "descripcion": {
                  "type": "string",
                  "example": "descripcion_2"
                },
                "nombre": {
                  "type": "string",
                  "example": "tarea_2"
                },
                "fechaInicio": {
                  "type": "string",
                  "example": "2020-03-02T03:00:00.000Z"
                },
                "fechaFin": {
                  "type": "string",
                  "example": "2020-01-04T03:00:00.000Z"
                },
                "numeroDeOrden": {
                  "type": "number",
                  "example": 2
                },
                "tipoDeTarea": {
                  "type": "object",
                  "properties": {
                    "nombre": {
                      "type": "string",
                      "example": "tarea3"
                    },
                    "descripcion": {
                      "type": "string",
                      "example": "descripcion_3"
                    },
                    "_id": {
                      "type": "string",
                      "example": "615e53c8db8fc5cc55e1fb42"
                    }
                  }
                },
                "sector": {
                  "type": "object",
                  "properties": {
                    "operario": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    },
                    "nombre": {
                      "type": "string",
                      "example": "deposito"
                    },
                    "_id": {
                      "type": "string",
                      "example": "615e53c8db8fc5cc55e1fb43"
                    }
                  }
                },
                "estado": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "_id": {
                  "type": "string",
                  "example": "615e53c8db8fc5cc55e1fb41"
                }
              }
            }
          },
          "estados": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "fechaInicio": {
                  "type": "string",
                  "example": "2021-09-20T03:00:00.000Z"
                },
                "fechaFin": {
                  "type": "string",
                  "example": "2021-10-05T03:00:00.000Z"
                },
                "observacion": {
                  "type": "string",
                  "example": "obs_5"
                },
                "tipoEstado": {
                  "type": "object",
                  "properties": {
                    "nombre": {
                      "type": "string",
                      "example": "Iniciada"
                    },
                    "descripcion": {
                      "type": "string",
                      "example": ""
                    },
                    "_id": {
                      "type": "string",
                      "example": "615e5b1e2968697488389e79"
                    }
                  }
                },
                "_id": {
                  "type": "string",
                  "example": "615e5b1e2968697488389e78"
                }
              }
            }
          }
        }
      }
    },
...
  }
}